[REF] Created single header file
     - All the functions needed are in the spine-raylib.h file
